problems merging CORE-4738
