+++ go/client/cmd_simplefs_copy.go
@@ -54,7 +54,7 @@ func (c *CmdSimpleFSCopy) Run() error {
 
 	c.G().Log.Debug("SimpleFSCopy (recursive: %v) to: %s", c.recurse, pathToString(c.dest))
 
-	destPaths, err := doSimpleFSPlatformGlob(c.G(), ctx, cli, c.src)
+	destPaths, err := doSimpleFSGlob(c.G(), ctx, cli, c.src)
 	if err != nil {
 		return err
 	}
